app/requirements.txt
# FastAPI 核心
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
email-validator==2.1.0

# 資料庫連接
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.28.0
pymongo==4.6.0
motor==3.3.1
redis==4.5.5
aiobotocore==2.6.0

# 身份驗證和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
itsdangerous==2.1.2

# 消息佇列
aio-pika==9.3.0

# 郵件通知
fastapi-mail==1.4.1

# 推播通知
firebase-admin==6.2.0

# 開發工具
pytest==7.4.3
httpx==0.25.1
black==23.10.1
isort==5.12.0

# 文件生成
fastapi-pagination==0.12.11

# 日誌處理
loguru==0.7.2

# CORS 支援
starlette==0.27.0

# GraphQL 支援
strawberry-graphql==0.211.1

# WebSocket 支援
websockets==12.0

# API 限流與保護
slowapi==0.1.8
fastapi-limiter==0.1.5 